QPIDJMS-204 Fix intermittent test failure
The remote close occurs so close to the open that it showed a small race
where the session did not yet know about the producer and as such the
remote close event could not be propagated to it.  Consumer code has the
same issue and is fixed here as well.