QPIDJMS-164: Provide OSGi metadata for qpid-jms-client
This change adds the maven-bundle-plugin to the parent project and
activates it in the qpid-jms-client project, generating OSGi meta data
for the project.

The Bundle-SymbolicName is overridden to "org.apache.qpid.jms.client".

The DynamicPackage-Import header is set to "*" since the JMS client can
actually create instances from all kind of classes.

Signed-off-by: Jens Reimann <jreimann@redhat.com>