QPIDJMS-232 Perform authentication on transport connect
Do the authentication on transport connect as opposed to waiting until
the Connection is first used to allow for early detection of a connect
failure due to do bad credentials.  Connection will now fail from the
ConnectionFactory createConnection call if the user cannot be
authenticated instead of waiting until the connection is used.