QPIDJMS-242 Creating a JMS resource can hang during failover reconnect
During failover the recovery attempts to recreate JMS resources that
were previously opened but recovers some resources that are still in the
process of being created lead to a double create effect on recover that
leads to the initial create call to hang.  Need to track state of the
JMS resources so that the failover recover can skip any tracked
instances that are either not yet opened or have already been closed.