QPIDJMS-245 Ensure that encoded values on the URI are handled properly
In some cases the parameters where being encoded or decoded twice or
even an unbalanced number of times leading to corruption of the original
values.  Refactor a bit to ensure that we encode values when we create
the URIs and decode when we parse out the options consistently.  Adds
some additional tests to cover some of these cases.