QPIDJMS-256 QPIDJMS-231 Better deal with deffered consumer close
Handle consumer close in TX better by moving the deferment handling to
the AmqpConsumer where close decide if it should hold until the TX has
completed and ensure that on deferred close that the consumer is drained
and prefetched messages are released back to the remote for redispatch.
Ensure that when a consumer is closed but deferred due to being in a
transaction or having pending delivered messages that are not yet ack'd
the client does not dispatch any more inbound messages to the consumer.