QPIDJMS-267 Failover discovery via connection properties
Add support for discovering new failover URIs bsaed on redirect
style Maps in a list provided in the Open frame of a remote peer.
Allow for addition, replacement or ignore of those discovered
hosts by the failover transport and provide an event point for
connection listeners to see the discovered resources.