QPIDJMS-283 In-doubt state on JmsTransactionInfo should be volatile
The in-doubt state on the transaction info should be volatile as it may
be modified on connection recovery without a lock when a commit or
rollback request is pending on the failover transport.  Add some
additional test support features from ActiveMQ to let the integration
tests run in repoest cycles.