QPIDJMS-283 Ensure new TX started whenever there is pipelined failure
When offline the failover provider allows rollback to complete normally
as the TX has been terminated already so the command doesn't need to be
held.  We need to ensure that when these states happen we still look and
begin a new TX which will wait for connection restoration to complete.