QPIDJMS-317 Make Transfer tag generation and caching more efficient
Improve the way transfer tags are generated to speed up the process and
also reduce the amount of garbage created in the process.  Also improve
the tag cache to reduce simplify the tag pool such that it doesn't
create any additional garbage on checkout or checkin and optimizes the
checkin process.