Issue QPIDJMS-400: Add missing OSGI resolution:=optional to native transports
The packages io.netty.channel.epoll and io.netty.channel.kqueue need
to be marked as optional since only one of them can be activated at a
time depending on the target platform of choice. The OSGI system fails
to activate the bundle that does not match the platform where the
client is deployed and fails the application activation that attempts
to use it.