QPIDJMS-418 Optimize usage of Symbol types to reduce overhead
The code is currently doing many Symbol.valueOf calls for each message
on send and receive where it should be using cached Symbol values to
avoid the overhead of the Symbol objects cache lookups and string
compares.  Clean up some other ares where we process Symbols to reduce
overhead on send and receive.