QPIDJMS-420 Use local tracking in AmqpConsumer for prefetch refill
Track the number of message that are currently in the prefetch buffer using
local counters that mesh well with the already added delivered message counts
to reduce access to synchronized code in the message queue.  Reduce code paths
where a message settlement might be missed or credit not replenished on error.