QPIDJMS-467 Provide client API stack traces in JMS Exceptions
Only throw ProviderException types from the provider layer and allow
for conversion to the correct JMS exception type in the JMS layer to
allow the stack traces in the thrown exceptions to map to which JMS
API method was called to trigger the error.  Thrown exceptions will
provide the cause for the error from the proivider layer from the
calling method in the JMS layer.