QPIDJMS-519 Correctly recover transaction after failover
Adds additional guards against allowing a send or acknowledge
to occur outside the bounds of a transaction when recovering a
connection following failover.  Adds tests to ensure that the
transaction becomes in-doubt whenever the state of recovery
cannot ensure that there will be no lost work or untransacted
work.  Test added to cover cases where this was previously seen.